backend/requirements/base.txt
celery==5.4.0
django==4.2.30
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
psycopg[binary]==3.2.13
redis==5.0.8
